refactor: update routes
