+++ src/components/dashboard/club/main/molecules/CalendarShortcut.tsx
@@ -1,10 +1,11 @@
 import Link from "next/link";
+import { CLUB_DASHBOARD_ENDPOINT } from "@/lib/constants";
 import { ChevronRight } from "@/assets/icons/chevron";
 
 export default function DashboardCalendarShortcut() {
   return (
     <Link
-      href={"/dashboard/manage/calendar"}
+      href={`${CLUB_DASHBOARD_ENDPOINT}/manage/calendar`}
       className="flex justify-between p-8 rounded-xl bg-gray-0 select-none"
     >
       <h3 className="h1 font-bold text-brand-black truncate">
