+++ src/components/dashboard/club/main/molecules/Overview.tsx
@@ -1,4 +1,5 @@
 import Link from "next/link";
+import { CLUB_DASHBOARD_ENDPOINT } from "@/lib/constants";
 
 export default function DashboardOverview() {
   return (
@@ -9,7 +10,7 @@ export default function DashboardOverview() {
           <span className="h4 font-medium text-gray-400">
             {"동호회 신규가입 신청"}
           </span>
-          <Link href={"/dashboard/manage/member"}>
+          <Link href={`${CLUB_DASHBOARD_ENDPOINT}/manage/member?filter=new`}>
             <span className="h1 font-extrabold text-brand-orange underline underline-offset-4 decoration-gray-800 hover:decoration-brand-orange transition duration-300">
               {`${0}건`}
             </span>
@@ -20,7 +21,7 @@ export default function DashboardOverview() {
           <span className="h4 font-medium text-gray-400">
             {"동호회 문의 접수"}
           </span>
-          <Link href={"/dashboard/manage?tab=qna"}>
+          <Link href={`${CLUB_DASHBOARD_ENDPOINT}/manage?tab=qna`}>
             <span className="h1 font-extrabold text-gray-0 underline underline-offset-4 decoration-gray-800 hover:decoration-gray-0 transition duration-300">{`${0}건`}</span>
           </Link>
         </div>
@@ -29,7 +30,7 @@ export default function DashboardOverview() {
           <span className="h4 font-medium text-gray-400">
             {"읽지 않은 인사 공지사항"}
           </span>
-          <Link href={"/dashboard/announcement"}>
+          <Link href={`${CLUB_DASHBOARD_ENDPOINT}/announcement?filter=company`}>
             <span className="h1 font-extrabold text-gray-0 underline underline-offset-4 decoration-gray-800 hover:decoration-gray-0 transition duration-300">{`${0}건`}</span>
           </Link>
         </div>
@@ -38,7 +39,7 @@ export default function DashboardOverview() {
           <span className="h4 font-medium text-gray-400">
             {"읽지 않은 omo 공지사항"}
           </span>
-          <Link href={"/dashboard/announcement"}>
+          <Link href={`${CLUB_DASHBOARD_ENDPOINT}/announcement?filter=omo`}>
             <span className="h1 font-extrabold text-gray-0 underline underline-offset-4 decoration-gray-800 hover:decoration-gray-0 transition duration-300">{`${0}건`}</span>
           </Link>
         </div>
