+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const CLUB_DASHBOARD_ENDPOINT = "/dashboard/club";
